backend/requirements.txt
fastapi[all]
diffusers==0.31.0
transformers[sentencepiece]
scipy
ftfy
accelerate
protobuf==3.20.1
celery[redis]
python-socketio[asyncio]
asgiref
